requirements.txt
python-telegram-bot
aiosqlite
pydantic
pydantic-settings
openai
httpx
gTTS
pydub
aiohttp
SpeechRecognition
